chore: add vitest and super basic snapshot test for adi-o-matic 3000 script
