+++ mock-data.csv
@@ -1,4 +1,4 @@
-color-token, foo
-color-token-of-doom,red-500
 color-token-of-greatness,blue-500
+color-token,foo
+color-token-of-doom,red-500
 color-button-primary-wombat-text,aqua-600
